LayoutTests/fast/dom/SelectorAPI/attrname-case-insensitive-expected.txt

PASS document.querySelector('div[baR]') is document.getElementById('bar')
PASS document.getElementById('bar').webkitMatchesSelector('div[baR]') is true
PASS document.querySelector('div[bar]') is document.getElementById('bar')
PASS document.getElementById('bar').webkitMatchesSelector('div[bar]') is true
PASS document.querySelector('div[BAR]') is document.getElementById('bar')
PASS document.getElementById('bar').webkitMatchesSelector('div[BAR]') is true
PASS document.querySelector('div[bAR]') is document.getElementById('bar')
PASS document.getElementById('bar').webkitMatchesSelector('div[bAR]') is true
PASS document.querySelector('div[baR="10"]') is document.getElementById('bar')
PASS document.getElementById('bar').webkitMatchesSelector('div[baR="10"]') is true
PASS document.querySelector('div[bar="10"]') is document.getElementById('bar')
PASS document.getElementById('bar').webkitMatchesSelector('div[bar="10"]') is true
PASS document.querySelector('div[BAR="10"]') is document.getElementById('bar')
PASS document.getElementById('bar').webkitMatchesSelector('div[BAR="10"]') is true
PASS document.querySelector('div[bAR="10"]') is document.getElementById('bar')
PASS document.getElementById('bar').webkitMatchesSelector('div[bAR="10"]') is true
PASS document.querySelector('path[pathLength]') is document.getElementById('pa')
PASS document.getElementById('pa').webkitMatchesSelector('path[pathLength]') is true
PASS content.querySelector('path[pathlength]') is null
PASS content.querySelector('path[pathLengTh]') is null
PASS document.querySelector('path[pathLength="200"]') is document.getElementById('pa')
PASS document.getElementById('pa').webkitMatchesSelector('path[pathLength="200"]') is true
PASS content.querySelector('path[pathlength="200"]') is null
PASS content.querySelector('path[pathLengTh="200"]') is null
PASS successfullyParsed is true

TEST COMPLETE
